pyproject.toml
[project]
name = "group-filmwork-watching-service"
version = "0.1.0"
description = "Service which allows to watch filmworks together"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mypy>=1.11.1",
    "alembic==1.13.2",
    "sqlalchemy==2.0.32",
    "pydantic-settings==2.4.0",
    "uvicorn>=0.30.6",
    "asyncpg>=0.29.0",
    "greenlet>=3.0.3",
    "fastapi[standard]>=0.112.2",
    "orjson>=3.10.7",
    "aiohttp>=3.10.5",
    "asgi-correlation-id>=4.3.2",
    "pyjwt>=2.9.0",
    "cryptography>=43.0.0",
    "redis>=5.0.8",
    "ruff>=0.6.2",
    "gunicorn>=23.0.0",
    "pytest>=8.3.2",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
lint.select = ["E", "F", "I", "UP", "B", "SIM",]
line-length = 120
